+++ MelonLoader/Fixes/XTermFix.cs
@@ -1,3 +1,4 @@
+#if !WINDOWS && !NET6_0_OR_GREATER
 using System;
 using System.Collections.Generic;
 using System.Linq;
@@ -115,3 +116,4 @@ internal static class XTermFix
         return list;
     }
 }
+#endif