+++ MelonLoader/Melons/RottenMelon.cs
@@ -12,6 +12,13 @@ namespace MelonLoader
         public readonly string errorMessage;
         public readonly Exception exception;
 
+        public RottenMelon(MelonAssembly assembly, string errorMessage, Exception exception = null)
+        {
+            this.assembly = assembly;
+            this.errorMessage = errorMessage;
+            this.exception = exception;
+        }
+
         public RottenMelon(Type type, string errorMessage, Exception exception = null)
         {
             assembly = MelonAssembly.LoadMelonAssembly(null, type.Assembly);